gui: Use localised time in duration
https://github.com/syncthing/syncthing/pull/8291 inpired me to
develop this. I tested it with all the languages Syncthing
currently supports, and they all work.

The only issue is that when you change the language in the gui,
you have to either refresh the page, or wait a few seconds, for
the page to refresh by itself, before the duration is translated
into the new language.