Fix timezone mismatch between WebAuthn registration and loadState
The time zone of CreateTime and LastUseTime was set by default during
registration and returned so to the GUI, but then when the state is read back
out of the DB these fields do not have a time zone set. This causes
`$scope.webauthnStateChanged` in the GUI to detect a false positive of pending
changes.