Implement changes requested by @calmh:
1. The build process creates a compatability.json
for the runtime used.
2. The release process adds compatability.json as
a release artifact
3. The upgrade server reads compatability.yaml
and merges it into its /meta.json response.
4. The upgrading client reads the
compatability.json file in the release bundle,
and using the go runtime setting found in it,
looks up the minimum OS versions listed in the
upgrade server's /meta.json response.
5. If the upgrading client's OS version is lower
than the minimum OS version listed, the upgrade
is aborted.