Add no-translate attribute for suppressing translation extraction
PR [#9175][1] currently adds the following element:

```html
<label>RP ID</label><!-- RP ID is the name of the WebAuthn setting, should not be translated -->
```

This is in a table of WebAuthn credentials that have been made ineligible
because they are bound to a different RP ID than the one currently configured in
Syncthing. "RP ID" is the technical term, and it probably should not be
translated since it relates to the "Webauthn Rp Id" setting in advanced
settings.

`go run build.go translate` prints a warning because this text is not
translated:

```
$ go run build.go translate
2024/08/21 21:58:14 Untranslated text node (settingsModalView.html):
2024/08/21 21:58:14     RP ID
```

But it shouldn't be either, so there should be a way to silence the warning for
when this is intentional.

This introduces a `no-translate` attribute that can be added to do just that.
This replaces the current special case for `id="contributor-list"` and
`id="copyright-notices"`, and deletes the latter attribute since it is not
referenced anywhere else.

An explanation can be added as the attribute value - for example:
`no-translate="Don't translate list of contributors"`. This value is printed if
`go run build.go translate` is run with the `TRANSLATEDEBUG` environment
variable set to non-empty:

```
$ TRANSLATEDEBUG=true go run build.go translate
2024/08/21 22:20:09 Ignoring no-translate <div> in aboutModalView.html: Don't translate list of contributors
2024/08/21 22:20:09 Ignoring no-translate <ul> in aboutModalView.html: Don't translate list of copyright notices
```

[1]: https://github.com/syncthing/syncthing/pull/9175