Separate "WebAuthn credentials" tables into separate scroll boxes
On narrow screens, the explanation row of the "Ineligible WebAuthn Credentials"
table balloons the width of both tables together, pulling the "Actions" column
of both tables way out to the far right. This change doesn't solve the issue of
the "Actions" column getting hidden even when "Ineligible WebAuthn Credentials"
is hidden, but at least decouples the scroll widths of the two.