lib/discovery: Backoff for global lookup failures
We have some scaling challenges around global discovery. These are made
significantly worse by Syncthing's retry behavior on failure. Normally a
"not found" can be cached for a while, but connection failures do not
result in negative cache entries. The result is that if the service is
down for a minute or two there are now tens or hundreds of thousands of devices
doing quite frequent retries, resulting in a massive thundering herd
when the service comes back up again.

This adds an increasing backoff for connection failures or unexpected
server errors.