Reduce WebAuthn config message redundancy
The message "Reload page at {%webauthnOrigin%} to enable WebAuthn." does not
need to be shown when only `isLocationInsecure()`, as this is already covered by
the combination of the message "WebAuthn requires HTTPS on non-localhost
domains." and the button "Reload page at {%webauthnOrigin%}". The former message
is still relevant when `!locationMatchesWebauthnOrigin()`, though.