Don't require versionBase in docsURL
In https://github.com/syncthing/syncthing/pull/9175 we will sometimes want to
show links to the documentation on the login page. These links currently get
truncated to just `https://docs.syncthing.net`, discarding the section path,
because the server version is not yet known while on the login page.

I don't think it's any worse to try to preserve the section path even without an
explicit version tag, than to fall back to just the host and lose all context
the link was attempting to provide.