fix(ur): actually send usage report directly when enabled
There was a bug that the unique ID was not set when reporting was
enabled, and thus the reports where rejected by the server. The unique
ID got set only on startup, so next time Syncthing restarted.

This makes sure to set the unique ID when blank. The check to send a new
report when the ID changes is removed, because it would result in double
reports being sent, and is in general not required for anything.