chore(gui): cache input type in each advanced settings category
Each section in the advanced settings dialog has similar code to
insert repeated input fields for each option.  But only the first
section (GUI options) was adjusted in #9743 to avoid calling the
inputTypeFor() function repeatedly.

Apply the same caching to a locally scoped variable for each ng-repeat
entry by defining it in an ng-init directive.