feat(log): Add finer grained logging
This PR allows one to set the STTRACE environment variable as:

  facility[:level][,facility2[:level2]] ...

For example:

  all:warn,fs:info,model

will log everything at the WARN level (so no INFO lines in the logs),
logs the fs facility at the INFO level (no DEBUG lines),
and logs the model facility at the DEBUG level.

Abbreviations are allowed, so

  all:w,fs:i,model

is the same as above.

The PR is backwards compatible with the existing logic.