fix(config): Skip GUI port probing for UNIX sockets.
When creating an initial default config, we usually probe for a free
TCP port.  But when a UNIX socket is specified via the STGUIADDRESS=
override or the --gui-address=unix:///... command line syntax, parsing
that option will fail during port probing.

The solution is to just skip the port probing when the address is
determined to specify something other than a TCP socket.