fix(scanner): continue walk after special files (fixes #9872)
We must skip unix sockets, fifos, etc when scanning as these are not
filetypes we can handle. Currently we return a "bug" error, which
results in the walk being aborted and the rest of the tree being
essentially invisible to Syncthing. Instead, just ignore these files
and continue onwards.

This might well be #9859 as well but I can't confirm.