chore(scanner): avoid scan failures and report if they happen
Currently we simply ignore those errors. And these are errors that don't
just affect one file, but they caused the entire scan to abort and thus
leaving potential a large amount of files unchecked. This shouldn't ever
happen, so log at warning level and add failure reporting.

When this came up, Jakob mentioned there are a bunch of error conditions
where we probably shouldn't actually abort. I agree, we should probably
never abort, even if the error is entirely unexpected - no harm trying
to continue to find more files that we may be able to scan. As such
streamline the handleItem resp. it's child `walk...` functions to simply
return an error, not doing anything special to categorize them. And in
the top-level walk function handle those errors without returning them.