chore(proto): change symlinktarget to be byte sequence
Strings in protobuf must be utf-8, and we enforce that explicitly or
implicitly in most places, but not really in symlink targets. Symlinks
can target things outside of synced folders and we have no authority to
decide what the encoding of that filename is, so allow it as an
uninterpreted byte sequence.

This is wire- and database compatible with the existing definition.