Delete untranslated translations
These seem obviously incorrect, and probably should not be present in the
translation files as they'll likely show up in tools as false positives.
Non-existent translation strings fall back to the English text automatically, so
there is no need to have the fallback in the translation files. Only if the
translation is identical to English (as is "OK", "LDAP" etc in almost all files)
should it be present in the translation file.

These changes were manually selected after generating candidate untranslations
using the following sed script:

    sed -i '/^\s*"\(.*\)": \?"\1",\?$/d' gui/default/assets/lang/lang-*.json