feat(ignore): Add .stignore escaping on Windows
Fixes #10057

Based on the discussion in
https://forum.syncthing.net/t/towards-syncthing-2-0/24072/35
This PR adds the ability for Windows users to use the pipe character (|) to
escape the metacharacters *, ?, [, and { in .stignore files.

Additionally, this PR adds the ability for the user to set the escape character
to backslash, or any character they want, by adding a line in the form:

 #escape=X

(where X is any single rune), to the top of the .stginore file.

This would allow users to use the same .stignore files across platforms, by
simply adding

 #escape=\

to their files.