Create an info gauge for folder_info
This includes the information most likely to make it possible for a
human to correlate the ids to something meaningful.

There were a couple of approaches possible, but adjusting
`metrics.go` seems like the least invasive and likely to cause
difficulty.  A function on `folder` would require expanding the
surface area of `prometheus` which feel worse than expanding
syncthing's own `config` type.