fix(model): Close fd immediately in copier
The `defer fd.Close()` used to be part of a function argument passed to
a method iterating over block map entries from the DB. Now it's part of
this very long-lived copier routine, i.e. we keep the file open until the
entire pull operation is done. Which may take a long while and involve a
lot of files. A heap profile was already reported in the forum that did
show this.