chore(db, model): Simplify per hash DB lookup in copier
In my opinion the main win of this change is in it's lower complexity
resp. fewer moving parts. It should also be faster as it only does one
query instead of two, but I have no idea if that's practically
relevant.

This also mirrors the v1 DB, where a block map key had the name
appended. Not that this is an argument for the change, it was mostly
reassuring me that I might not be missing something key here
conceptually.

This came up while working on another bug in the copier (open file
leak). And the nested nature of this code makes the fix harder, and
"un-nesting" it required me to understand what's happening. This change
fell out of that.